Rename mutt/ctype.h to mutt/ctype2.h.
