+++ complete/helpers.c
@@ -44,7 +44,7 @@
 #include "menu/lib.h"
 #include "compapi.h"
 #include "data.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 
 /**
  * matches_ensure_morespace - Allocate more space for auto-completion
