+++ email/parse.c
@@ -48,7 +48,7 @@
 #include "from.h"
 #include "globals.h"
 #include "mime.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 #include "parameter.h"
 #include "rfc2047.h"
 #include "rfc2231.h"
