+++ email/rfc2047.c
@@ -42,7 +42,7 @@
 #include "rfc2047.h"
 #include "envelope.h"
 #include "mime.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 
 #define ENCWORD_LEN_MAX 75
 #define ENCWORD_LEN_MIN 9 /* strlen ("=?.?.?.?=") */
