+++ imap/command.c
@@ -53,7 +53,7 @@
 #include "edata.h"
 #include "mdata.h"
 #include "msn.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 #include "mutt_logging.h"
 #include "mx.h"
