+++ imap/imap.c
@@ -67,7 +67,7 @@
 #include "mdata.h"
 #include "msg_set.h"
 #include "msn.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 #include "mutt_logging.h"
 #include "mutt_socket.h"
 #include "muttlib.h"
