+++ key/lib.c
@@ -36,7 +36,7 @@
 #include "gui/lib.h"
 #include "key/lib.h"
 #include "menu/lib.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 
 extern const struct MenuFuncOp OpAlias[];
 extern const struct MenuFuncOp OpAttachment[];
