+++ mutt_config.c
@@ -44,7 +44,7 @@
 #include "attach/lib.h"
 #include "expando/lib.h"
 #include "index/lib.h"
-#include "mutt/ctype.h"
+#include "mutt/ctype2.h"
 #include "mutt_logging.h"
 #include "mutt_thread.h"
 #include "mx.h"
